+++ CHANGELOG.md
@@ -7,12 +7,12 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
-- Add new boolean configuration `allow-local-imports` to allow for local imports
-
 ### Enhancements
 
+- Add new boolean configuration `allow-local-imports` to allow for local imports
 - Include the pycodestyle error code to the error message for PEP8 style errors
 - Added date and time display to `PlainReporter` and `ColorReporter`
+- Allowed specifying allowed names in configurations `allowed-import-modules` and `extra-imports` instead of just modules
 
 ## [2.7.0] - 2024-12-14